[CREATE]: create fetchRegistrationResponse function for register
